carbon_tracker_agent/requirements.txt
streamlit
pandas
plotly
python-dotenv
google-generativeai
reportlab
statsmodels
